Allow custom function for checking if tab is active
Sometimes $state.include is not enough for cheking whether a parent tab is active, for example when the parent tab links to a default child tab.  In thiks case when navigating to another child tab would cause the parent to become deselected.  This allows a custom function to be defined in the tabs object which will be called to determine whether a tab is selected.  If the function is note present the default (current) function will be used.